[InstantSearch result page] Refactored the folder organisation
