+++ .gitignore
@@ -3,3 +3,4 @@ css/.sass-cache/
 .DS_Store
 node_modules/*
 npm-debug.log
+node_modules
