--- instant-search/instantsearch.js/index.html
@@ -1,71 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <title>Instant Search Tutorial</title>
-  <meta http-equiv="X-UA-Compatible" content="IE=Edge">
-  <!-- IE < 10 support -->
-  <!--[if lte IE 9]>
-    <script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
-  <![endif]-->
-  <link rel="stylesheet" href="//cdn.jsdelivr.net/instantsearch.js/1/instantsearch.min.css">
-  <link rel="stylesheet" type="text/css" href="style.css">
-</head>
-<body>
-
-  <!-- INTERFACE -->
-  <header>
-    <a href="."><img src="img/instant_search_logo@2x.png"/></a>
-    <div id="search-input"></div>
-    <div id="search-input-icon"></div>
-  </header>
-
-  <main>
-    <div id="left-column">
-      <div id="category" class="facet"></div>
-      <div id="brand" class="facet"></div>
-      <div id="price" class="facet"></div>
-      <div id="type" class="facet"></div>
-    </div>
-
-    <div id="right-column">
-      <div id="sort-by-wrapper"><span id="sort-by"></span></div>
-      <div id="stats"></div>
-      <div id="hits"></div>
-      <div id="pagination"></div>
-    </div>
-  </main>
-
-  <footer>
-    Source Code on <a href="https://github.com/algolia/examples/blob/master/instant-search/instantsearch.js">Github</a> -
-    Powered by <a href="https://www.algolia.com">Algolia</a> - Data from <a href="https://developer.bestbuy.com">Best Buy</a>
-  </footer>
-  <!-- /INTERFACE -->
-
-  <!-- TEMPLATES -->
-  <script type="text/html" id="hit-template">
-    <div class="hit">
-      <div class="hit-image">
-        <img src="{{image}}" alt="{{name}}">
-      </div>
-      <div class="hit-content">
-        <h3 class="hit-price">${{price}}</h3>
-        <h2 class="hit-name">{{{_highlightResult.name.value}}}</h2>
-        <p class="hit-description">{{{_highlightResult.description.value}}}</p>
-      </div>
-    </div>
-  </script>
-
-  <script type="text/html" id="no-results-template">
-    <div id="no-results-message">
-      <p>We didn't find any results for the search <em>"{{query}}"</em>.</p>
-      <a href="." class='clear-all'>Clear search</a>
-    </div>
-  </script>
-  <!-- /TEMPLATES -->
-
-  <!-- SCRIPTS -->
-  <script src="//cdn.jsdelivr.net/instantsearch.js/1/instantsearch.min.js"></script>
-  <script src="app.js"></script>
-  <!-- /SCRIPTS -->
-</body>
-</html>
